feat: move walletModal to separate feature
