+++ features/header/headerWallet.tsx
@@ -3,8 +3,8 @@ import { CHAINS, getChainColor } from '@lido-sdk/constants';
 import { useSDK } from '@lido-sdk/react';
 import { useWeb3 } from '@reef-knot/web3-react';
 import { ThemeToggler } from '@lidofinance/lido-ui';
-import { WalletButton, WalletConnect } from 'features/wallet';
 import { HeaderWalletChainStyle } from './headerWalletStyles';
+import { WalletButton, WalletConnect } from 'features/walletModal';
 
 export const HeaderWallet: FC = () => {
   const { active } = useWeb3();
