+++ features/snapshot/snapshotForm.tsx
@@ -7,7 +7,7 @@ import { InputAddress, addressValidator } from 'shared/ui/inputAddress';
 import { useEncodeSnapshotCalldata } from 'features/votingAdapter';
 import { Form } from './snapshotFormStyles';
 import { useWeb3 } from 'reef-knot';
-import { WalletConnect } from 'features/wallet';
+import { WalletConnect } from 'features/walletModal';
 
 type SnapshotFormData = {
   delegateAddress: string;
