+++ features/wallet/index.ts
@@ -1,7 +1,4 @@
 export * from './fallbackWallet';
-export * from './providers';
 export * from './wallet';
-export * from './walletButton';
 export * from './walletCard';
-export * from './walletConnect';
-export * from './walletModal';
+export * from './web3Provider';
