--- features/wallet/providers/index.tsx
@@ -1,3 +1,0 @@
-export * from './modals';
-export * from './web3';
-export * from './walletProviders';
