--- features/wallet/walletButton/index.ts
@@ -1,1 +1,0 @@
-export * from './walletButton';
