--- features/wallet/walletCard/index.ts
@@ -1,1 +1,0 @@
-export * from './walletCard';
