--- features/wallet/walletConnect/index.ts
@@ -1,1 +1,0 @@
-export * from './walletConnect';
