--- features/wallet/walletModal/index.ts
@@ -1,1 +1,0 @@
-export * from './walletModal';
