+++ pages/_app.tsx
@@ -7,11 +7,12 @@ import {
   migrationThemeCookiesToCrossDomainCookiesClientSide,
   CookieThemeProvider,
 } from '@lidofinance/lido-ui';
-import { WalletProviders } from 'features/wallet';
+import { Web3Provider } from 'features/wallet';
 import { withCsp } from 'shared/api/withCsp';
 import { GlobalStyle } from 'shared/ui';
 import { ClaimingProvider } from 'features/claim';
 import { VestingsProvider } from 'features/vesting';
+import { ModalProvider } from 'features/walletModal';
 
 // Migrations old cookies to new cross domain cookies
 migrationThemeCookiesToCrossDomainCookiesClientSide();
@@ -35,16 +36,18 @@ const AppWrapper = (props: AppProps): JSX.Element => {
 
   return (
     <CookieThemeProvider>
-      <GlobalStyle />
-      <WalletProviders>
-        <ClaimingProvider>
-          <VestingsProvider>
-            <MemoApp {...rest} />
-            <CookiesTooltip />
-            <ToastContainer />
-          </VestingsProvider>
-        </ClaimingProvider>
-      </WalletProviders>
+      <Web3Provider>
+        <ModalProvider>
+          <GlobalStyle />
+          <ClaimingProvider>
+            <VestingsProvider>
+              <MemoApp {...rest} />
+              <CookiesTooltip />
+              <ToastContainer />
+            </VestingsProvider>
+          </ClaimingProvider>
+        </ModalProvider>
+      </Web3Provider>
     </CookieThemeProvider>
   );
 };
