+++ shared/ui/index.ts
@@ -3,6 +3,7 @@ export * from './backgroundGradient';
 export * from './footer';
 export * from './formatDate';
 export * from './formatToken';
+export * from './inputAmount';
 export * from './inputGroupStyled';
 export * from './inputNumber';
 export * from './main';
